Generate Traditional Chinese locale at build time
