+++ package.json
@@ -9,6 +9,7 @@
     "lint": "eslint . --ext js,jsx,cjs,mjs,ts,tsx --fix",
     "guard:app-jsx": "node scripts/guard-app-jsx-edits.mjs",
     "guard:encoding": "node scripts/guard-encoding.mjs",
+    "locale:zh-tw": "node scripts/generate-zh-tw-locale.mjs",
     "test:unit": "node --test test/unit/*.test.mjs",
     "verify:release": "node scripts/verify-release.mjs",
     "postinstall": "cross-env PYTHONUTF8=1 electron-builder install-app-deps",
@@ -118,7 +119,6 @@
     "marked": "^18.0.0",
     "music-metadata": "^11.12.3",
     "naudiodon": "^2.3.6",
-    "opencc-js": "^1.3.0",
     "qrcode": "^1.5.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
@@ -136,6 +136,7 @@
     "electron-vite": "^2.3.0",
     "eslint": "^8.57.1",
     "eslint-plugin-react": "^7.37.4",
+    "opencc-js": "^1.3.0",
     "prettier": "^3.4.2",
     "vite": "^5.4.14"
   },
